Moved world settings into the settings application directory.
Renamed sheet settings content to client settings for honesty's sake.
