+++ src/applications/settings/client-settings/tabs/GmOptionsSettingsTab.svelte
@@ -6,7 +6,7 @@
   import { SettingsProvider } from 'src/settings/settings';
   import TextInputSetting from 'src/applications/settings/parts/TextInputSetting.svelte';
   import SelectSetting from 'src/applications/settings/parts/SelectSetting.svelte';
-  import type { SettingsSheetContext } from '../SheetSettings.types';
+  import type { SettingsSheetContext } from '../ClientSettings.types';
 
   let context = getContext<Writable<SettingsSheetContext>>('context');
