+++ src/applications/settings/client-settings/tabs/InfoTab.svelte
