+++ src/applications/settings/client-settings/tabs/PlayerSettingsTab.svelte
@@ -10,7 +10,7 @@
   import type {
     SettingsSheetContext,
     SettingsSheetFunctions,
-  } from '../SheetSettings.types';
+  } from '../ClientSettings.types';
   import ListboxSetting from '../parts/ListboxSetting.svelte';
   import { CONSTANTS } from 'src/constants';
